frontend/requirements.txt
streamlit==1.11.0
plotly==5.8.2
pandas==1.4.1
rich==12.4.4
pretty-errors==1.2.25
kafka-python==2.0.2